Move polling code in its own module
It's time to bite the bullet and move this code that has nothing to do
with tachyon in its own module.
