+++ test/teiserver_web/tachyon/session_test.exs
@@ -2,7 +2,7 @@ defmodule TeiserverWeb.Tachyon.SessionTest do
   use TeiserverWeb.ConnCase
   alias WebsocketSyncClient, as: WSC
   alias Teiserver.Support.Tachyon
-  import Teiserver.Support.Tachyon, only: [poll_until_some: 1, poll_until: 2]
+  import Teiserver.Support.Polling, only: [poll_until_some: 1, poll_until: 2]
   alias Teiserver.Player
 
   setup _context do
@@ -57,7 +57,7 @@ defmodule TeiserverWeb.Tachyon.SessionTest do
     sess_pid = Player.SessionRegistry.lookup(user.id)
     assert sess_pid != nil
     ref = Process.monitor(sess_pid)
-    Tachyon.poll_until(fn -> Player.lookup_connection(user.id) end, &is_nil/1)
+    poll_until(fn -> Player.lookup_connection(user.id) end, &is_nil/1)
     send(sess_pid, :player_timeout)
     assert_receive({:DOWN, ^ref, :process, _, _}, 1000, "Session should have died")
   end
